# next-app/openapi.yaml
...
openapi: 3.0.3
info:
  title: QueroUmaCor API (Next.js)
  version: 1.0.0
  description: |
    Backend endpoints for the Next.js port of the QueroUmaCor PWA. Runs on
    the edge runtime (`runtime = 'edge'`). Most endpoints are 1:1 ports of
    the legacy Cloudflare Pages Functions in `functions/api/*`; see
    `next-app/lib/api/_services/*` for the business logic and the vanilla
    `openapi.yaml` at repo root for the legacy contract.
servers:
  - url: https://app2.queroumacor.com.br/api
    description: Staging (Next.js preview on Cloudflare Pages)
  - url: https://queroumacor.com.br/api
    description: Production (cutover target — currently still served by vanilla)
...
paths:
...
  /me-export:
    post:
      tags: [profile]
      summary: LGPD Art. 18 V — export ALL personal data for the logged-in user.
      description: |
        Returns a downloadable JSON file (Content-Disposition: attachment)
        with the user's profile, quotes, orders, messages, notifications,
        notes, points, reviews, follows, likes, comments, posts and
        referrals. Rate limit: 3/min per user (expensive query fan-out).
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                  description: Optional fallback when no Authorization header.
      responses:
        '200':
          description: Personal data dump.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="queroumacor-meus-dados-abcd1234.json"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: See _meta.schema_version and per-table arrays.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { description: Internal error. }
        '502': { description: Failed to validate token. }
        '503': { description: Service role not configured. }
...
  /chat-ai:
    post:
      tags: [ai]
      summary: Conversation with Seu Ze (the mascot assistant).
      description: |
        Conversational helper for painters. PRO-gated, rate limit 20/min.
        OpenAI primary, Gemini fallback.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                accessToken: { type: string }
                message: { type: string, maxLength: 1500 }
                history:
                  type: array
                  maxItems: 10
                  items:
                    type: object
                    properties:
                      role: { type: string, enum: [user, assistant] }
                      content: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: AI reply.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
        '400': { description: Invalid JSON / missing message. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/ProRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { description: Internal error. }
        '502': { description: All providers failed. }
        '503': { description: No AI keys configured. }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT (access_token from supabase-js). Several endpoints
        also accept the token via `accessToken` field in the JSON body
        or multipart FormData (vanilla-client compat) when the
        Authorization header is inconvenient.
  responses:
    Unauthenticated:
      description: Missing or invalid token / login required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ProRequired:
      description: Endpoint requires `profile.is_pro = true`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded for this endpoint+user.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              limit: { type: integer }
              count: { type: integer }
              retry_after_seconds: { type: integer }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...